move public getIconFromMimeType to another module
